requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.10
pydantic==1.10.7
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
psycopg2-binary==2.9.6
redis==4.5.4
alembic==1.10.3
python-dotenv==0.21.1
